requirements.txt
langchain==0.1.0
pydub==0.25.1
python-dotenv==1.0.0
pytube==15.0.0
psutil==5.9.7
keyboard==0.13.5
streamlit==1.29.0
streamlit_extras==0.3.6
openai==1.7.0
chromadb==0.4.22
tiktoken==0.5.2
langchain_openai==0.0.2
yt_dlp==2023.12.30